feat(svc-account-auth): Fixes required to support service account authentication
Closes VC-31642
