+++ pkg/venafi/cloud/connectorSSH.go
@@ -1,0 +1,19 @@
+package cloud
+
+import "github.com/Venafi/vcert/v5/pkg/certificate"
+
+func (c *Connector) RetrieveSshConfig(_ *certificate.SshCaTemplateRequest) (*certificate.SshConfig, error) {
+	panic("operation is not supported yet")
+}
+
+func (c *Connector) RetrieveSSHCertificate(_ *certificate.SshCertRequest) (response *certificate.SshCertificateObject, err error) {
+	panic("operation is not supported yet")
+}
+
+func (c *Connector) RequestSSHCertificate(_ *certificate.SshCertRequest) (response *certificate.SshCertificateObject, err error) {
+	panic("operation is not supported yet")
+}
+
+func (c *Connector) RetrieveAvailableSSHTemplates() (response []certificate.SshAvaliableTemplate, err error) {
+	panic("operation is not supported yet")
+}
