+++ pkg/venafi/cloud/connector_test.go
@@ -1176,7 +1176,7 @@ func TestSetPolicy(t *testing.T) {
 	}
 
 	//validate each attribute
-	userDetails, err := getUserDetails(conn)
+	userDetails, err := conn.getUserDetails()
 	//validating the default users attribute was created
 	users := []string{
 		//"jenkins@opensource.qa.venafi.io",
